total calcs 2.0

+++ src/Admin.java
@@ -10,7 +10,7 @@
 public class Admin extends User
 {
     private String username, password, address;
-    int coupon;
+    double coupon;
     Admin()
     {
         firstName = "Bob";
